Made changes to functioning of Light Wand
Added variable to control the max charge, power per use, and power
restore of the Light Wand. In addtion also decreased the max charge,
power per use, and power restore of the Light Wand. Also changed the
Light Wand so that when it is out of energy it can still return Heated
Stone nodes into Stone in order to restore energy.